requirements.txt
fastapi>=0.100.0,<0.110.0
uvicorn>=0.22.0,<0.29.0
openai>=1.3.5,<1.25.0
pydantic>=2.1,<2.6
pydantic-settings>=2.0,<2.2
python-jose[cryptography]>=3.3.0,<4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pytest>=7.0,<9.0
pytest-asyncio>=0.20,<0.24
black>=23.3,<25.0
slowapi>=0.1.4,<0.2.0
locust>=2.15,<3.0
httpx<0.26
